some simple refactor. want to define interfaces in definition files, will do later.
